Update README and test-plugin template for correct language content generation
The README file has been updated to show the problem of the site generating content in the correct language, but failing in other languages. It has been translated to another language. Also, changes were made in the test-plugin template to ensure content is loaded in the correct language. In both cases, changes aim to improve multilingual site generation.